Add cross-links to links.json (#168)
Closes https://github.com/elastic/docs-builder/issues/162

## Context

For our Links Validation Service, we need to generate a list of
cross-links (links pointing to external documentation in our system)
that will be uploaded to the Link Index.

## Changes

Add external links  to the links.json

Resulting JSON file:

```json
{
  "origin": {
    "branch": "feature/add-external-links",
    "remote": "elastic/docs-builder-unknown",
    "ref": "03c22a29e6585b72f64a0aa64f087dd5664c939f"
  },
  "url_path_prefix": "",
  "internal_links": [
    "index.md",
    "configure/index.md",
    "configure/page.md",
    "testing/external-links.md",
    "...",
  ],
  "cross_links": [
    "kibana://index.md",
    "elasticsearch://index.md"
  ]
}
```

## Notes

Changed the naming to Cross-Links, so it won't get confused with
external hosts.