Support anchors and table of contents from included files (#585)
* Support anchors and table of contents from included files

If an included snippet defines it's own headers and anchors they did not contribute to the pages anchors and headers.

-- page.md
```markdown
:::{include} _snippet/snippet.md
:::
```

This resulted in link check failures to `page.md#included-from-snippet`, as well as the headers from a snippet not appearing on the page's table of contents.

This PR also extends our testing framework by allowing to collect data from the conversion using `IConversionCollector`.

The `authoring` test framework uses the same `DocumentationGenerator` that `docs-builder` uses. The default for `IConversionCollector` is `null` in real world usage as we don't want the added memory pressure.

An upside of this is that we can now really fully test the whole HTML layout vs just the markdown HTML.

Lastly this PR includes a tiny fix to DiagnosticLinkInlineParser.cs to always report the resolved path as the full path. That way folks don't need to parse all the parent up `../` instructions themselves when dealing with this error instance.

* Add back AngleSharp.Diffing

* ensure we lookup snippets anchored to the root

* fix license headers

* don't lookup anchors in includes that are not found (or are cyclical)

* remove htmx tags from test assertions for now