Separate static url prefix from path prefix (#744)
* Separate static url prefix from path prefix

It still defaults to UrlPathPrefix but with the assembler build we need a way to ensure static files can
be rooted seperately from the path prefix of the page.

To help with development we now have a 'dev' environment which defaults to http://localhost:4000/docs

We can now view the output of `docs-assembler repo build all` with `docs-builder serve-static`
which will host the combined docs under http://localhost:4000

* remove unused namespaces

* Adjust .editorconfig for StaticWebHost diagnostic suppression

Added explicit diagnostic suppression rules for IL3050 and IL2026 under the StaticWebHost.cs section in the .editorconfig file. This ensures consistent handling of these diagnostics specifically for StaticWebHost-related code.